+++ src/main/ui/src/components/Recharts/Recharts.css
