+++ src/main/ui/src/interno/relatorio_vaga/relatorioVaga.css
