--- usersapi/src/main/ui/src/interno/home/Componentes/sidebar/Sidebar.css
@@ -1,147 +1,0 @@
-
-.sidebar {
-  background-color: var(--sidebar-bg-color);
-  width: var(--sidebar-width);
-  transition: all 0.3s ease-in-out;
-  transform: translateX(0);
-  z-index: 9999999;
-}
-
-.sidebar .close-aside {
-  position: absolute;
-  top: 7px;
-  right: 7px;
-  cursor: pointer;
-  color: #eee;
-}
-
-.sidebar .sidebar-header {
-  border-bottom: 1px solid #ffffff;
-}
-
-.sidebar .sidebar-header h5 a {
-  color: var(--dk-gray-300);
-}
-
-.sidebar .sidebar-header p {
-  color: var(--dk-gray-400);
-  font-size: 0.825rem;
-}
-
-.sidebar .search .form-control ~ i {
-  color: #2b2f3a;
-  right: 40px;
-  top: 22px;
-}
-
-.sidebar > ul > li {
-  padding: 0.7rem 1.75rem;
-}
-
-.sidebar ul > li > a {
-  color: #d1d5db;
-  text-decoration: none;
-}
-
-/* Start numbers */
-.sidebar ul > li > a > .num {
-  line-height: 0;
-  border-radius: 3px;
-  font-size: 14px;
-  padding: 0px 5px;
-}
-
-.sidebar ul > li > i {
-  font-size: 18px;
-  margin-right: 0.7rem;
-  color: var(--dk-gray-500);
-}
-
-.sidebar ul > li.has-dropdown > a:after {
-  content: "\eb3a";
-  font-family: unicons-line;
-  font-size: 1rem;
-  line-height: 1.8;
-  float: right;
-  color: var(--dk-gray-500);
-  transition: all 0.3s ease-in-out;
-}
-
-.sidebar ul .opened > a:after {
-  transform: rotate(-90deg);
-}
-
-/* Start dropdown menu */
-.sidebar ul .sidebar-dropdown {
-  padding-top: 10px;
-  padding-left: 30px;
-  display: none;
-}
-.sidebar ul .sidebar-dropdown.active {
-  display: block;
-}
-
-.sidebar ul .sidebar-dropdown > li > a {
-  font-size: 0.85rem;
-  padding: 0.5rem 0;
-  display: block;
-}
-
-.btnSideBar{
-  padding: 0.5rem 1rem;
-  border-radius: 5px;
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.btnSideBar img{
-  height: 1.5rem;
-}
-
-@media screen and (max-width:  1024px) and (min-width: 769px){
-  .sidebar > ul > li{
-    padding: 0;
-  }
-
-  .margem-relatorio {
-    margin: 1% 20px 0 calc(var(--sidebar-width) + 2%);
-  }
-}
-
-@media screen and (max-width:  768px) and (min-width: 501px) {
-  .categories li a span{
-    display: none;
-  }
-
-  .sidebar-header{
-    flex-direction: column;
-  }
-
-  .margem-relatorio {
-    margin: 1% 20px 0 calc(var(--sidebar-width) + 2.5%);
-  }
-
-}
-
-@media screen and (max-width:  500px) {
-  .btnSideBar{
-      padding: .5rem 23px;
-  }
-
-  .categories li a span{
-    display: none;
-  }
-
-  .sidebar-header{
-    flex-direction: column;
-  }
-
-  .sidebar > ul > li{
-    padding: 0;
-  }
-
-  .margem-relatorio {
-    margin: 1% 20px 0 calc(var(--sidebar-width) + 3.5%);
-  }
-}
